drivers:platform:aducm3029: i2c read fix
Setting bRepeatedStart=true for a transaction does not only ommit the stop
bit of the current transaction, but actually sends the repeated start as
well as a prologue data sequence.
This means that calling a read without a
stop bit and thn a write for aducm3029 actually sends the reapeated start,
then the I2C address and then a Stop bit directly, meaning the write
operation is not atomic anymore.
At best this is a benign bug and might still work, with hardware errors and
reading wrong registers at worst.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>